feat(draco): the

+++ core/src/main/java/com/mndk/bteterrarenderer/draco/metadata/AttributeMetadata.java
@@ -12,7 +12,7 @@ import lombok.EqualsAndHashCode;
 @EqualsAndHashCode(callSuper = true)
 public class AttributeMetadata extends Metadata {
 
-    private int attUniqueId;
+    private long attUniqueId;
 
     public AttributeMetadata() {
         super();
